+++ src/LiveComponent/tests/Functional/Metadata/LiveComponentMetadataFactoryTest.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\UX\LiveComponent\Tests\Functional\Metadata;
+
+use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
+use Symfony\UX\LiveComponent\Metadata\LiveComponentMetadataFactory;
+use Symfony\UX\LiveComponent\Tests\Fixtures\Component\ComponentWithUrlBoundProps;
+
+class LiveComponentMetadataFactoryTest extends KernelTestCase
+{
+    public function testQueryStringMapping()
+    {
+        /** @var LiveComponentMetadataFactory $metadataFactory */
+        $metadataFactory = self::getContainer()->get('ux.live_component.metadata_factory');
+
+        $class = new \ReflectionClass(ComponentWithUrlBoundProps::class);
+        $propsMetadata = $metadataFactory->createPropMetadatas($class);
+
+        $propsMetadataByName = [];
+        foreach ($propsMetadata as $propMetadata) {
+            $propsMetadataByName[$propMetadata->getName()] = $propMetadata;
+        }
+
+        $this->assertTrue($propsMetadataByName['prop1']->queryStringMapping());
+
+        $this->assertTrue($propsMetadataByName['prop2']->queryStringMapping());
+
+        $this->assertTrue($propsMetadataByName['prop3']->queryStringMapping());
+
+        $this->assertFalse($propsMetadataByName['prop4']->queryStringMapping());
+
+        $this->assertTrue($propsMetadataByName['prop5']->queryStringMapping());
+
+        $this->assertTrue($propsMetadataByName['prop6']->queryStringMapping());
+    }
+}